// client/src/domain/model/product/product.type.ts
export type ProductType = {
    _id:  string;
    sku:  string;
    barcode:  string;
    name: string;
    description: string;
    images: string[];
    category: string;
    type: string;
    brand: string;
    color: string;
    model: string;
    gender: string;
    size: string;
    grossPrice: number;
    stock: number;
    active:boolean;
};